move errors (since public API)
